Add files via upload
Those are tests to serial communication with the Arduino's USB port. they work only once, if the program stop any loop or end up stopping, it will require an USB reset (reconnect the USB on port) so it can reset the permission, since i'm on windows and still don't have lots of experience in serial communication, i would be glad if some one could make this more feasible, or how to make the permission never change in the port. Otherwise, the program will not allow you to compile neither the .ino nor the .py. making both program to stop each other.